backend/django_app/requirements.txt
djangorestframework 
django-cors-headers
django
psycopg2-binary  
requests 